--- a/sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/encoders/RowEncoderSuite.scala
+++ b/sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/encoders/RowEncoderSuite.scala
@@ -17,6 +17,7 @@
 
 package org.apache.spark.sql.catalyst.encoders
 
+import scala.collection.immutable
 import scala.collection.mutable
 import scala.util.Random
 
@@ -317,8 +318,8 @@ class RowEncoderSuite extends CodegenInterpretedPlanTest {
     val schema = new StructType().add("a", ArrayType(dt, nullable))
     test(s"RowEncoder should return mutable.ArraySeq with properly typed array for $schema") {
       val encoder = ExpressionEncoder(schema).resolveAndBind()
-      val result = fromRow(encoder, toRow(encoder, Row(data))).getAs[mutable.ArraySeq[_]](0)
-      assert(result.array.getClass === data.getClass)
+      val result = fromRow(encoder, toRow(encoder, Row(data))).getAs[immutable.ArraySeq[_]](0)
+      assert(result.unsafeArray.getClass === data.getClass)
       assert(result === data)
     }
   }