### What changes were proposed in this pull request?
The primary change in this PR is in the function `RowEncoder#encoderForDataType`, where the encoding type for `ArrayType` data has been changed from `immutable.ArraySeq[_]` to `immutable.ArraySeq[_]` in order to optimize the performance of `Row#getSeq` to match the performance when using Spark 3.5 with Scala 2.12.

Following the modification of `RowEncoder.scala`, the data transferred between the JVM and R may now include `immutable.ArraySeq`. As a result, this pr also fix `r/SerDe.scala` to adjust to this scenario and guarantees that all existing test cases can pass successfully.

The changes to other test files are because, after this pr, Seq type data will be encoded as `immutable.ArraySeq` type, so it is necessary to use `getAs[immutable.Seq]` instead of `getAs[mutable.Seq]` to retrieve them, but if `getSeq` is used to retrieve the data, there will be no changes.

### Why are the changes needed?
Optimize the performance of `Row#getSeq` to match the performance when using Spark 3.5 with Scala 2.12.

Consider the following scenario:

```scala
    val data = (0 until arraySize).toArray
    val row = Seq(data).toDF().collect().head

    val benchmark = new Benchmark(
      s"Test get seq with $arraySize from row",
      valuesPerIteration,
      output = output)

    benchmark.addCase("Get Seq") { _: Int =>

      for (_ <- 0L until valuesPerIteration) {
        val ret = row.getSeq(0)
      }
    }

    benchmark.run()
```

Run tests on the aforementioned code with `arraySize` set to 10, 100, 1000, 10000, and 100000, respectively, and with `valuesPerIteration` set to 100000, getting the following results:

- branch-3.5(with Scala 2.12):

```
OpenJDK 64-Bit Server VM 1.8.0_422-b05 on Linux 5.15.0-1068-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 10 from row:            Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               1              1           0        194.8           5.1       1.0X

OpenJDK 64-Bit Server VM 1.8.0_422-b05 on Linux 5.15.0-1068-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 100 from row:           Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               1              1           0         96.8          10.3       1.0X

OpenJDK 64-Bit Server VM 1.8.0_422-b05 on Linux 5.15.0-1068-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 1000 from row:          Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               1              1           0         97.0          10.3       1.0X

OpenJDK 64-Bit Server VM 1.8.0_422-b05 on Linux 5.15.0-1068-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 10000 from row:         Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               1              1           0         96.8          10.3       1.0X

OpenJDK 64-Bit Server VM 1.8.0_422-b05 on Linux 5.15.0-1068-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 100000 from row:        Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               1              1           0         96.9          10.3       1.0X

```

- master (with Scala 2.13):

```
OpenJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 10 from row:            Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               9             10           0         10.5          94.8       1.0X

OpenJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 100 from row:           Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                              65             65           1          1.5         646.4       1.0X

OpenJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 1000 from row:          Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                             614            615           1          0.2        6140.2       1.0X

OpenJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 10000 from row:         Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                            6122           6128           8          0.0       61223.1       1.0X

OpenJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 100000 from row:        Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                           61247          61268          30          0.0      612468.1       1.0X
```

We can observe that in branch-3.5(with Scala 2.12), the latency of `Row#getSeq` is constant, whereas in master(with Scala 2.13), the latency of `Row#getSeq` exhibits a linearly increasing trend with the length of the array.

This is due to the following code:

https://github.com/apache/spark/blob/c121bb557447709c19fc01b68b3a34b8838abc6d/sql/api/src/main/scala/org/apache/spark/sql/Row.scala#L317-L325

The default Scala version in branch-3.5 is 2.12, where `Seq` represents `scala.collection.Seq`. Therefore, the `res.toSeq` in the code above is a redundant operation that does not trigger a collection copy.

However, the default Scala version in master is 2.13, where `Seq` represents `scala.collection.immutable.Seq`, and `getAs[scala.collection.Seq[T]](i)` returns a `scala.collection.mutable.ArraySeq[T]` type. As a result, the `res.toSeq` in the code above triggers a collection copy from `mutable.ArraySeq[T]` to `immutable.ArraySeq[T]`.

This has caused a performance regression for the `Row#getSeq` API between master and Spark 3.5 with Scala 2.12


### Does this PR introduce _any_ user-facing change?
Yes.

If users utilize the `Row#getSeq` API to fetch data, there will be no user-facing change.

However, if users previously used the `Row#getAs[mutable.Seq]` API to obtain data, they will need to change it to `Row#getAs[immutable.Seq]`.

### How was this patch tested?
- Pass GitHub Actions
- Test the previously described scenario with this PR:


```
penJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 10 from row:            Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               0              0           0        338.5           3.0       1.0X

OpenJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 100 from row:           Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               0              0           0        338.2           3.0       1.0X

OpenJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 1000 from row:          Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               0              0           0        337.4           3.0       1.0X

OpenJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 10000 from row:         Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               0              0           0        338.1           3.0       1.0X

OpenJDK 64-Bit Server VM 17.0.12+7-LTS on Linux 6.5.0-1025-azure
AMD EPYC 7763 64-Core Processor
Test get seq with 100000 from row:        Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
Get Seq                                               0              0           0        337.9           3.0       1.0X
```

We can see that, with this PR, the latency of the `Row#getSeq` API has returned to being constant.


### Was this patch authored or co-authored using generative AI tooling?
No