@LuciferYang 2024-08-09
Another possible fix is to changet the return type of `Row#getSeq`:

Change the code from

https://github.com/apache/spark/blob/af70aafd330fdbb6ce0d5b3efbcb180cda488695/sql/api/src/main/scala/org/apache/spark/sql/Row.scala#L317-L325

to

```
def getSeq[T](i: Int): scala.collection.Seq[T] = getAs[scala.collection.Seq[T]](i)
```

However, this would lead to a failure in the MiMa check:

```
[error] spark-sql-api: Failed binary compatibility check against org.apache.spark:spark-sql-api_2.13:3.5.0! Found 1 potential problems (filtered 327)
[error]  * method getSeq(Int)scala.collection.immutable.Seq in interface org.apache.spark.sql.Row does not have a correspondent in current version
[error]    filter with: ProblemFilters.exclude[DirectMissingMethodProblem]("org.apache.spark.sql.Row.getSeq")

``

Therefore, I did not choose this approach.