@LuciferYang 2024-08-09
> So, the following two main class are the target?
> 
> * core/src/main/scala/org/apache/spark/api/r/SerDe.scala
> * sql/api/src/main/scala/org/apache/spark/sql/catalyst/encoders/RowEncoder.scala

Yes, the changes to `RowEncoder.scala` were made to optimize performance. 
After the changes to `RowEncoder.scala`, there will be `immutable.ArraySeq` in the data passed between the JVM and R, so it is also necessary to fix `r/SerDe.scala`, otherwise some SparkR-related tests will fail.