@LuciferYang 2024-08-09
> Thank you, @LuciferYang , although the title could be misleading because `master` and `Apache Spark 3.5.2 (Scala 2.13)` will be the same. :)

I've made some changes to the pr title; do you think it's a little better now? Thanks