@LuciferYang 2024-08-10
> Thank you. So much better. BTW, you know, we don't test `[TESTS]` for the PR touching `src/main`.
> 
> * core/src/main/scala/org/apache/spark/api/r/SerDe.scala
> * sql/api/src/main/scala/org/apache/spark/sql/catalyst/encoders/RowEncoder.scala
> 
> If we start to tag `[TESTS]` for all PRs containing unit tests, `[TESTS]` itself becomes meaningless because all PRs are supposed to have a unit test. Could you remove it from the PR title?

Thank you for pointing out this.