@LuciferYang 2024-08-10
> According to the test case change and this comment, you mean this is a breaking change, right?
> 
> > But if the user is used to using row.getAs[mutable.Seq] instead of row.getSeq, then he/she will have to switch to using row.getAs[immutable.Seq].

Yes, for this behavioral habit, it is a breaking change.