@LuciferYang 2024-08-10
@dongjoon-hyun I made another attempt: https://github.com/apache/spark/pull/47698, and I believe this should also solve the problem without introducing a breaking change.